Criando detais do gastos

+++ src/main/java/br/com/skeleton/spendsmart/resource/response/ExpenseDetailResponse.java
@@ -17,7 +17,7 @@ public class ExpenseDetailResponse {
 
     private String name;
 
-    private Double value;
+    private BigDecimal value;
 
     private InstallmentDetailResponse installmentDetailResponse;
